Send quantity amounts as strings
If we convert quantities, which we get as strings, to floats, we lose
some precision, which manifests itself as unintended additional digits
when Wikibase converts the float back to a decimal number: a quantity
originally sent as 3.14 can end up being saved on the entity as
3.140000000000000124344978758017532527446746826171875 instead. It’s
difficult to avoid this for the upper and lower bounds, since we need to
add the error to the amount, but we can at least avoid it for the amount
itself, by just sending the original string to Wikibase.

Credit for the investigation of this issue, including the example values
cited above, goes to Matěj Suchánek, who also proposed an alternative
solution in D911 (which also has more discussion).