Fix links to forms and senses in old interface
Forms and senses do not have pages of their own, but are rather sections
of the general lexeme page. Since T195477, it is fortunately very easy
to convert form and sense IDs to (namespace-less) titles: simply replace
the dash in their ID with a hash, making the form- or sense-specific
part of the ID the fragment.