Fix links to entities in new interface
Instead of assuming that properties live in the Property: namespace and
everything else (assumption: only items; not true with lexicographical
data) lives in the main namespace, use the map of letters to namespace
prefixes we already have in the config. This requires a companion change
in magnustools to actually make use of the map we’re setting up here.